+++ packages/e2e-tests/tests/common/create-sdk.ts
@@ -4,14 +4,12 @@ import type { BlockchainWallet } from "@rarible/sdk-wallet"
 import type { RaribleSdkEnvironment } from "@rarible/sdk/src/config/domain"
 import { createRaribleSdk } from "@rarible/sdk"
 import { LogsLevel } from "@rarible/sdk/build/domain"
+import { testsConfig } from "./config"
 
 export function createSdk(blockchain: Blockchain, wallet: BlockchainWallet): IRaribleSdk {
-	let env: RaribleSdkEnvironment = "e2e"
+	let env: RaribleSdkEnvironment = testsConfig.env as RaribleSdkEnvironment
 
 	switch (blockchain) {
-		case Blockchain.TEZOS:
-			env = "dev" // @todo: tezos not working in e2e, need to fix (?)
-			break
 		case Blockchain.POLYGON:
 			env = "dev"
 			break