+++ packages/example/src/pages/bid/bid-prepare-form.tsx
@@ -1,0 +1,58 @@
+import React, { useContext } from "react"
+import { Box, Stack } from "@mui/material"
+import { useForm } from "react-hook-form"
+import { toItemId } from "@rarible/types"
+import { PrepareBidResponse } from "@rarible/sdk/build/types/order/bid/domain"
+import { faChevronRight } from "@fortawesome/free-solid-svg-icons"
+import { FormTextInput } from "../../components/common/form/form-text-input"
+import { FormSubmit } from "../../components/common/form/form-submit"
+import { resultToState, useRequestResult } from "../../components/hooks/use-request-result"
+import { ConnectorContext } from "../../components/connector/sdk-connection-provider"
+import { RequestResult } from "../../components/common/request-result"
+
+interface IBidPrepareFormProps {
+	disabled?: boolean
+	onComplete: (response: PrepareBidResponse) => void
+}
+
+export function BidPrepareForm({ disabled, onComplete }: IBidPrepareFormProps) {
+	const connection = useContext(ConnectorContext)
+	const form = useForm()
+	const { handleSubmit } = form
+	const { result, setError } = useRequestResult()
+
+	return (
+		<>
+
+			<form onSubmit={handleSubmit(async (formData) => {
+				if (!connection.sdk) {
+					return
+				}
+				try {
+					onComplete(await connection.sdk.order.bid({
+						itemId: toItemId(formData.itemId)
+					}))
+				} catch (e) {
+					setError(e)
+				}
+			})}
+			>
+				<Stack spacing={2}>
+					<FormTextInput form={form} name="itemId" label="Item ID"/>
+					<Box>
+						<FormSubmit
+							form={form}
+							label="Next"
+							state={resultToState(result.type)}
+							icon={faChevronRight}
+							disabled={disabled}
+						/>
+					</Box>
+				</Stack>
+			</form>
+			<Box sx={{ my: 2 }}>
+				<RequestResult result={result}/>
+			</Box>
+		</>
+	)
+}