+++ packages/sdk/src/sdk-blockchains/ethereum/bid.test.ts
@@ -1,13 +1,16 @@
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import { EthereumWallet } from "@rarible/sdk-wallet"
 import { awaitAll, deployTestErc1155, deployTestErc20, deployTestErc721 } from "@rarible/ethereum-sdk-test-common"
-import { toAddress, toContractAddress, toItemId } from "@rarible/types"
+import { toAddress, toContractAddress, toCurrencyId, toItemId } from "@rarible/types"
 import BigNumber from "bignumber.js"
+import type { EthErc20AssetType } from "@rarible/api-client"
 import { Blockchain } from "@rarible/api-client"
 import { createRaribleSdk as createEtherumSdk } from "@rarible/protocol-ethereum-sdk"
+import { sentTx } from "@rarible/protocol-ethereum-sdk/build/common/send-transaction"
 import { createRaribleSdk } from "../../index"
 import { retry } from "../../common/retry"
 import { LogsLevel } from "../../domain"
+import { MintType } from "../../types/nft/mint/domain"
 import { initProvider, initProviders } from "./test/init-providers"
 import { awaitItem } from "./test/await-item"
 import { awaitStock } from "./test/await-stock"
@@ -27,24 +30,30 @@ describe("bid", () => {
 	})
 
 	const ethereum1 = new Web3Ethereum({ web3: web31 })
-	const ethwallet1 = new EthereumWallet(ethereum1, Blockchain.ETHEREUM)
-	const sdk1 = createRaribleSdk(ethwallet1, "e2e", { logs: LogsLevel.DISABLED })
+	const ethwallet1 = new EthereumWallet(ethereum1)
+	const sdk1 = createRaribleSdk(ethwallet1, "development", { logs: LogsLevel.DISABLED })
 
 	const ethereum2 = new Web3Ethereum({ web3: web32 })
-	const ethwallet2 = new EthereumWallet(ethereum2, Blockchain.ETHEREUM)
-	const sdk2 = createRaribleSdk(ethwallet2, "e2e")
-	const ethSdk2 = createEtherumSdk(ethwallet2.ethereum as any, "e2e", { logs: LogsLevel.DISABLED })
+	const ethwallet2 = new EthereumWallet(ethereum2)
+	const sdk2 = createRaribleSdk(ethwallet2, "development", { logs: LogsLevel.DISABLED })
+	const ethSdk2 = createEtherumSdk(ethwallet2.ethereum as any, "dev-ethereum", { logs: LogsLevel.DISABLED })
 
-	const { web3 } = initProvider()
+
+	const { web3 } = initProvider(undefined, {
+		rpcUrl: "https://dev-ethereum-node.rarible.com",
+		networkId: 300500,
+	})
 	const nullFundsEthereum = new Web3Ethereum({ web3: web3 })
-	const nullFundsWallet = new EthereumWallet(nullFundsEthereum, Blockchain.ETHEREUM)
-	const nullFundsSdk = createRaribleSdk(nullFundsWallet, "e2e", { logs: LogsLevel.DISABLED })
+	const nullFundsWallet = new EthereumWallet(nullFundsEthereum)
+	const nullFundsSdk = createRaribleSdk(nullFundsWallet, "development", { logs: LogsLevel.DISABLED })
 
-	const wethContractEthereum = toAddress("0xc6f33b62a94939e52e1b074c4ac1a801b869fdb2")
+	const wethContractEthereum = toAddress("0x55eB2809896aB7414706AaCDde63e3BBb26e0BC6")
 	const wethContract = toContractAddress(`${Blockchain.ETHEREUM}:${wethContractEthereum}`)
 	const wethAsset = { "@type": "ERC20" as const, contract: wethContract }
 
-	const e2eErc1155V2ContractAddress = convertEthereumContractAddress("0x268dF35c389Aa9e1ce0cd83CF8E5752b607dE90d", Blockchain.ETHEREUM)
+	const erc721Address = toAddress("0x96CE5b00c75e28d7b15F25eA392Cbb513ce1DE9E")
+	const erc1155Address = toAddress("0xda75B20cCFf4F86d2E8Ef00Da61A166edb7a233a")
+	const e2eErc1155V2ContractAddress = convertEthereumContractAddress(erc1155Address, Blockchain.ETHEREUM)
 
 	const it = awaitAll({
 		testErc20: deployTestErc20(web31, "Test1", "TST1"),
@@ -52,28 +61,40 @@ describe("bid", () => {
 		testErc1155: deployTestErc1155(web31, "Test2"),
 	})
 
-	test("bid on erc721 <-> erc20 and update bid", async () => {
+	test.skip("bid on erc721 <-> erc20 and update bid", async () => {
 		const itemOwner = await ethwallet1.ethereum.getFrom()
 
 		const bidderAddress = await ethwallet2.ethereum.getFrom()
 		const bidderUnionAddress = convertEthereumToUnionAddress(bidderAddress, Blockchain.ETHEREUM)
 
-		const tokenId = "1"
-		const itemId = convertEthereumItemId(`${it.testErc721.options.address}:${tokenId}`, Blockchain.ETHEREUM)
-		await it.testErc721.methods.mint(itemOwner, tokenId, "123").send({
-			from: itemOwner,
+		await sentTx(it.testErc20.methods.mint(bidderAddress, "10000000000000"), {
+			from: await ethereum1.getFrom(),
 			gas: 500000,
 		})
-		await it.testErc20.methods.mint(bidderAddress, "10000000000000").send({
-			from: itemOwner,
-			gas: 500000,
+
+		const action = await sdk1.nft.mint({
+			collectionId: convertEthereumContractAddress(erc721Address, Blockchain.ETHEREUM),
 		})
-		await awaitItem(sdk1, itemId)
+		const result = await action.submit({
+			uri: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG5",
+			creators: [{
+				account: convertEthereumToUnionAddress(itemOwner, Blockchain.ETHEREUM),
+				value: 10000,
+			}],
+			royalties: [],
+			lazyMint: false,
+			supply: 1,
+		})
+		if (result.type === MintType.ON_CHAIN) {
+			await result.transaction.wait()
+		}
+
+		await awaitItem(sdk1, result.itemId)
 
 		await resetWethFunds(ethwallet2, ethSdk2, wethContractEthereum)
 
-		const response = await sdk2.order.bid({ itemId })
-		const price = "0.00000000000000002"
+		const response = await sdk2.order.bid({ itemId: result.itemId })
+		const price = "0.0000000000000002"
 		const orderId = await response.submit({
 			amount: 1,
 			price,
@@ -93,7 +114,7 @@ describe("bid", () => {
 		const updateAction = await sdk2.order.bidUpdate({
 			orderId,
 		})
-		await updateAction.submit({ price: "0.00000000000000004" })
+		await updateAction.submit({ price: "0.0000000000000004" })
 
 		const acceptBidResponse = await sdk1.order.acceptBid({ orderId })
 		const acceptBidTx = await acceptBidResponse.submit({ amount: 1, infiniteApproval: true })
@@ -101,7 +122,7 @@ describe("bid", () => {
 
 		await retry(10, 1000, async () => {
 			return sdk1.apis.ownership.getOwnershipById({
-				ownershipId: `ETHEREUM:${it.testErc721.options.address}:${tokenId}:${bidderAddress}`,
+				ownershipId: `${result.itemId}:${bidderAddress}`,
 			})
 		})
 	})
@@ -161,13 +182,13 @@ describe("bid", () => {
 
 		const tokenId = "3"
 		const itemId = convertEthereumItemId(`${it.testErc1155.options.address}:${tokenId}`, Blockchain.ETHEREUM)
-		await it.testErc1155.methods.mint(senderRaw, tokenId, 100, "123").send({
+		await sentTx(it.testErc1155.methods.mint(senderRaw, tokenId, 100, "123"), {
 			from: senderRaw,
 			gas: 500000,
 		})
 		const bidderAddress = await ethereum2.getFrom()
 		const bidderUnionAddress = convertEthereumToUnionAddress(bidderAddress, Blockchain.ETHEREUM)
-		await it.testErc20.methods.mint(bidderAddress, "10000000000000").send({
+		await sentTx(it.testErc20.methods.mint(bidderAddress, "10000000000000"), {
 			from: senderRaw,
 			gas: 500000,
 		})
@@ -194,6 +215,16 @@ describe("bid", () => {
 
 			expect(value).toBe(undefined)
 		})
+		const value = await bidResponse.getConvertableValue({
+			currencyId: toCurrencyId(`ETHEREUM:${it.testErc20.options.address}`),
+			price: "0.00000000000000001",
+			amount: 5,
+			originFees: [{
+				account: bidderUnionAddress,
+				value: 1000,
+			}],
+		})
+		expect(value).toBe(undefined)
 	})
 
 	test("getConvertValue returns insufficient type", async () => {
@@ -201,14 +232,13 @@ describe("bid", () => {
 
 		const tokenId = "4"
 		const itemId = convertEthereumItemId(`${it.testErc1155.options.address}:${tokenId}`, Blockchain.ETHEREUM)
-		await it.testErc1155.methods.mint(senderRaw, tokenId, 100, "123").send({
+		await sentTx(it.testErc1155.methods.mint(senderRaw, tokenId, 100, "123"), {
 			from: senderRaw,
 			gas: 500000,
 		})
 		await awaitItem(nullFundsSdk, itemId)
 
 		const bidResponse = await nullFundsSdk.order.bid({ itemId })
-
 		await retry(5, 2000, async () => {
 			const value = await bidResponse.getConvertableValue({
 				assetType: { "@type": "ERC20", contract: wethContract },
@@ -224,10 +254,22 @@ describe("bid", () => {
 			expect(value.value.toString()).toBe("0.000000000000000055")
 			expect(value.type).toBe("insufficient")
 		})
+		const value = await bidResponse.getConvertableValue({
+			currencyId: toCurrencyId(wethContract),
+			price: "0.00000000000000001",
+			amount: 5,
+			originFees: [{
+				account: convertEthereumToUnionAddress(await ethereum2.getFrom(), Blockchain.ETHEREUM),
+				value: 1000,
+			}],
+		})
+		if (!value) throw new Error("Convertable value must be non-undefined")
+		expect(value.value.toString()).toBe("0.000000000000000055")
+		expect(value.type).toBe("insufficient")
 	})
 
 	test("getConvertableValue returns undefined", async () => {
-		const itemId = toItemId(`${Blockchain.ETHEREUM}:0xF04881F205644925596Fee9D66DACd98A9b99F05:1`)
+		const itemId = toItemId(`${Blockchain.ETHEREUM}:0x2Ac19979c171F7b626096C9eDc8Cd5C589cf110b:1`)
 
 		const bidResponse = await sdk2.order.bid({ itemId })
 
@@ -235,7 +277,6 @@ describe("bid", () => {
 		const wethAsset = { "@type": "ERC20" as const, contract: wethContract }
 		const wethBidderBalance = new BigNumber(await sdk2.balances.getBalance(bidderUnionAddress, wethAsset))
 
-		console.log("wethBidderBalance", wethBidderBalance.toString())
 		if (wethBidderBalance.lt("0.000000000000001")) {
 			const tx = await ethSdk2.balances.convert(
 				{ assetClass: "ETH" },
@@ -261,7 +302,7 @@ describe("bid", () => {
 	})
 
 	test("getConvertableValue returns convertable value", async () => {
-		const itemId = toItemId(`${Blockchain.ETHEREUM}:0xF04881F205644925596Fee9D66DACd98A9b99F05:1`)
+		const itemId = toItemId(`${Blockchain.ETHEREUM}:0x2Ac19979c171F7b626096C9eDc8Cd5C589cf110b:1`)
 
 		const bidResponse = await sdk2.order.bid({ itemId })
 
@@ -289,14 +330,14 @@ describe("bid", () => {
 		const ownerCollectionAddress = await ethereum1.getFrom()
 		const bidderAddress = await ethereum2.getFrom()
 
-		await it.testErc20.methods.mint(bidderAddress, "10000000000000").send({
-      	from: ownerCollectionAddress,
-      	gas: 500000,
+		await sentTx(it.testErc20.methods.mint(bidderAddress, "10000000000000"), {
+			from: ownerCollectionAddress,
+			gas: 500000,
 		})
 		const tokenId = "5"
 		const itemId = convertEthereumItemId(`${it.testErc721.options.address}:${tokenId}`, Blockchain.ETHEREUM)
 
-		await it.testErc721.methods.mint(ownerCollectionAddress, tokenId, "0").send({
+		await sentTx(it.testErc721.methods.mint(ownerCollectionAddress, tokenId, "0"), {
 			from: ownerCollectionAddress,
 			gas: 500000,
 		})
@@ -315,6 +356,7 @@ describe("bid", () => {
 				"@type": "ERC20",
 				contract: erc20Contract,
 			},
+			expirationDate: new Date(Date.now() + 60000),
 		})
 
 		const acceptBidResponse = await sdk1.order.acceptBid({
@@ -328,13 +370,114 @@ describe("bid", () => {
 		await fillBidResult.wait()
 	})
 
+	test("bid on erc721 <-> erc20 with CurrencyId", async () => {
+		const itemOwner = await ethwallet1.ethereum.getFrom()
+
+		const bidderAddress = await ethwallet2.ethereum.getFrom()
+		const bidderUnionAddress = convertEthereumToUnionAddress(bidderAddress, Blockchain.ETHEREUM)
+
+		const tokenId = "7"
+		const itemId = convertEthereumItemId(`${it.testErc721.options.address}:${tokenId}`, Blockchain.ETHEREUM)
+		await sentTx(it.testErc721.methods.mint(itemOwner, tokenId, "123"), {
+			from: itemOwner,
+			gas: 500000,
+		})
+		await sentTx(it.testErc20.methods.mint(bidderAddress, "10000000000000"), {
+			from: itemOwner,
+			gas: 500000,
+		})
+		await awaitItem(sdk1, itemId)
+
+		await resetWethFunds(ethwallet2, ethSdk2, wethContractEthereum)
+
+		const response = await sdk2.order.bid({ itemId })
+		const price = "0.00000000000000002"
+		const erc20Contract = convertEthereumContractAddress(it.testErc20.options.address, Blockchain.ETHEREUM)
+		const orderId = await response.submit({
+			amount: 1,
+			price,
+			currency: toCurrencyId(erc20Contract),
+			originFees: [{
+				account: bidderUnionAddress,
+				value: 1000,
+			}],
+		})
+
+		const order = await awaitStock(sdk1, orderId, price)
+		expect(order.makeStock.toString()).toEqual(price)
+		const takeAssetType = order.make.type as EthErc20AssetType
+		expect(takeAssetType["@type"]).toEqual("ERC20")
+		expect(takeAssetType.contract.toLowerCase()).toEqual(erc20Contract.toLowerCase())
+
+		const acceptBidResponse = await sdk1.order.acceptBid({ orderId })
+		const acceptBidTx = await acceptBidResponse.submit({ amount: 1, infiniteApproval: true })
+		await acceptBidTx.wait()
+
+		await retry(10, 1000, async () => {
+			return sdk1.apis.ownership.getOwnershipById({
+				ownershipId: `ETHEREUM:${it.testErc721.options.address}:${tokenId}:${bidderAddress}`,
+			})
+		})
+	})
+
+
+	test.skip("bid for collection with outdated expiration date", async () => {
+		const ownerCollectionAddress = await ethereum1.getFrom()
+		const bidderAddress = await ethereum2.getFrom()
+
+		await sentTx(it.testErc20.methods.mint(bidderAddress, "10000000000000"), {
+			from: ownerCollectionAddress,
+			gas: 500000,
+		})
+		const tokenId = "6"
+		const itemId = convertEthereumItemId(`${it.testErc721.options.address}:${tokenId}`, Blockchain.ETHEREUM)
+
+		await sentTx(it.testErc721.methods.mint(ownerCollectionAddress, tokenId, "1"), {
+			from: ownerCollectionAddress,
+			gas: 500000,
+		})
+		await awaitItem(sdk1, itemId)
+
+		const erc721Contract = convertEthereumContractAddress(it.testErc721.options.address, Blockchain.ETHEREUM)
+		const bidResponse = await sdk2.order.bid({
+			collectionId: erc721Contract,
+		})
+
+		const erc20Contract = convertEthereumContractAddress(it.testErc20.options.address, Blockchain.ETHEREUM)
+		const bidOrderId = await bidResponse.submit({
+			amount: 1,
+			price: "0.00000000000000001",
+			currency: {
+				"@type": "ERC20",
+				contract: erc20Contract,
+			},
+			expirationDate: new Date(),
+		})
+
+		const acceptBidResponse = await sdk1.order.acceptBid({
+			orderId: bidOrderId,
+		})
+		let errorMessage
+		try {
+			const fillBidResult = await acceptBidResponse.submit({
+				amount: 1,
+				infiniteApproval: true,
+				itemId: toItemId(`${erc721Contract}:${tokenId}`),
+			})
+			await fillBidResult.wait()
+		} catch (e: any) {
+			errorMessage = e.message
+		}
+		expect(errorMessage).toBeTruthy()
+	})
+
 	test.skip("bid for collection and accept bid on lazy item", async () => {
 		const ownerCollectionAddress = await ethereum1.getFrom()
 		const bidderAddress = await ethereum2.getFrom()
 
-		await it.testErc20.methods.mint(bidderAddress, "10000000000000").send({
-      	from: ownerCollectionAddress,
-      	gas: 500000,
+		await sentTx(it.testErc20.methods.mint(bidderAddress, "10000000000000"), {
+			from: ownerCollectionAddress,
+			gas: 500000,
 		})
 
 		const action = await sdk1.nft.mint({ collectionId: e2eErc1155V2ContractAddress })
@@ -349,6 +492,9 @@ describe("bid", () => {
 			lazyMint: true,
 			supply: 10,
 		})
+		if (mintResult.type === MintType.ON_CHAIN) {
+			await mintResult.transaction.wait()
+		}
 
 		await awaitItem(sdk1, mintResult.itemId)
 
@@ -365,8 +511,6 @@ describe("bid", () => {
 				contract: erc20Contract,
 			},
 		})
-
-		console.log("after bid")
 
 		const acceptBidResponse = await sdk1.order.acceptBid({
 			orderId: bidOrderId,