+++ packages/sdk/src/sdk-blockchains/flow/balance.test.ts
@@ -1,6 +1,6 @@
 import { FlowWallet } from "@rarible/sdk-wallet"
 import * as fcl from "@onflow/fcl"
-import { toContractAddress } from "@rarible/types"
+import { toContractAddress, toCurrencyId } from "@rarible/types"
 import { createApisSdk } from "../../common/apis"
 import { convertFlowUnionAddress } from "./common/converters"
 import { createFlowSdk } from "./index"
@@ -14,7 +14,7 @@ describe("Test flow balance function", () => {
 	const wallet = new FlowWallet(fcl)
 	const sdk = createFlowSdk(wallet, createApisSdk("prod"), "mainnet")
 
-	test("Should get FT balance for account", async () => {
+	test.skip("Should get FT balance for account", async () => {
 		const balance1 = await sdk.balances.getBalance(address, {
 			"@type": "FLOW_FT",
 			contract: toContractAddress("FLOW:A.0x1654653399040a61.FlowToken"),
@@ -25,6 +25,20 @@ describe("Test flow balance function", () => {
 			"@type": "FLOW_FT",
 			contract: toContractAddress("FLOW:A.0x3c5959b568896393.FUSD"),
 		})
+		expect(balance2.toString()).toEqual("0")
+	})
+
+	test("Should get FT balance for account with CurrencyId", async () => {
+		const balance1 = await sdk.balances.getBalance(
+			address,
+			toCurrencyId("FLOW:A.0x1654653399040a61.FlowToken")
+		)
+		expect(balance1.toString()).toEqual("0.001")
+
+		const balance2 = await sdk.balances.getBalance(
+			address,
+			toCurrencyId("FLOW:A.0x3c5959b568896393.FUSD")
+		)
 		expect(balance2.toString()).toEqual("0")
 	})
 })